// frontend/src/app/models/reservation.ts
export class Reservation {
    user: string = '';
    restourant: string = '';
    date: string = '';
    time: string = '';
    numOfPersons: string = ''
    additionalInfo: string = ''
    restourantAddress: string = ''
    status: string = ''
    declineComment: string = ''
    waiter: string = ''
    table: string = ''
  }